requirements.txt
streamlit
requests
azure-storage-blob
azure-identity
python-dotenv
numpy
pandas